+++ src/pytorch_lightning/CHANGELOG.md
@@ -94,6 +94,16 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 - Deprecated the functions in `pytorch_lightning.utilities.apply_func` in favor of `lightning_utilities.core.apply_func` ([#14516](https://github.com/Lightning-AI/lightning/pull/14516), [#14537](https://github.com/Lightning-AI/lightning/pull/14537))
 
 
+- Deprecated the `pytorch_lightning.plugins.io.CheckpointIO` in favor of `lightning_lite.plugins.io.CheckpointIO` ([#14519](https://github.com/Lightning-AI/lightning/pull/14519))
+
+
+- Deprecated the `pytorch_lightning.plugins.io.TorchCheckpointIO` in favor of `lightning_lite.plugins.io.TorchCheckpointIO` ([#14519](https://github.com/Lightning-AI/lightning/pull/14519))
+
+
+- Deprecated the `pytorch_lightning.plugins.io.XLACheckpointIO` in favor of `lightning_lite.plugins.io.XLACheckpointIO` ([#14519](https://github.com/Lightning-AI/lightning/pull/14519))
+
+
+
 ### Removed
 
 - Removed the deprecated `Trainer.training_type_plugin` property in favor of `Trainer.strategy` ([#14011](https://github.com/Lightning-AI/lightning/pull/14011))